Fix getDomainURL to return domain conf URLs
Previously it would return the domain provider base URL, which wasn't
what it was supposed to do. The function name was misleading so
that's probably where the confusion came from. This method is supposed
to return the configuration URL for a given domain you have signed up
with your (in this case Njalla) domain provider. For Njalla, such a
URL will be `https://njal.la/domains/mydomain.com/`, for instance.