Add record type structs with custom unmarshal
This is done so we get the benefits of type-checking on each Record
type. Each Record can, and has, slightly different arguments, so it's
good to have different structs for each differing type. But this means
we needed a custom unmarshal technique to be able to differentiate at
run-time what type of Record each was, and assign it to the proper
ones.

With this commits I've also added tests for example usage.